requirements/requirements.txt
django>=1.8.4,<1.9
django-password-validation==0.1.1
django-grappelli>=2.7.1,<2.8
python-memcached==1.59
mysql-python
pillow

# need an unreleased SNI fix from httplib2 lib TODO: remove after it is released to PyPi
-e git+https://github.com/httplib2/httplib2.git@5ccd260911e2db802fa97498ae92558fac8c8057#egg=httplib2-0.9.2_dev

-e git+https://github.com/open-craft/ansible-sanity-checker@v0.0.2#egg=sanity_checker
-e git+https://github.com/tophatmonocle/ims_lti_py.git@979244d83c2e6420d2c1941f58e52f641c56ad12#egg=ims_lti_py-develop
-e git+https://github.com/open-craft/django-lti-tool-provider@v0.1.3#egg=django_lti_tool_provider-master
-e git+https://github.com/edx/opaque-keys@27dc382ea587483b1e3889a3d19cbd90b9023a06#egg=opaque-keys
